Move masked input logic to the hook
