+++ docs/prop-types.json
@@ -845,7 +845,7 @@
     },
     "renderInput": {
       "defaultValue": null,
-      "description": "Override input component",
+      "description": "Render input component. Where `props` – [TextField](https://material-ui.com/api/text-field/#textfield-api) component props\n@example ```jsx\nrenderInput={props => <TextField {...props} />}\n````",
       "name": "renderInput",
       "parent": {
         "fileName": "material-ui-pickers/lib/src/_shared/PureDateInput.tsx",
@@ -912,10 +912,10 @@
         "name": "Partial<InputAdornmentProps>"
       }
     },
-    "KeyboardButtonProps": {
+    "OpenPickerButtonProps": {
       "defaultValue": null,
       "description": "Props to pass to keyboard adornment button",
-      "name": "KeyboardButtonProps",
+      "name": "OpenPickerButtonProps",
       "parent": {
         "fileName": "material-ui-pickers/lib/src/_shared/PureDateInput.tsx",
         "name": "DateInputProps"
@@ -1621,7 +1621,7 @@
     },
     "renderInput": {
       "defaultValue": null,
-      "description": "Override input component",
+      "description": "Render input component. Where `props` – [TextField](https://material-ui.com/api/text-field/#textfield-api) component props\n@example ```jsx\nrenderInput={props => <TextField {...props} />}\n````",
       "name": "renderInput",
       "parent": {
         "fileName": "material-ui-pickers/lib/src/_shared/PureDateInput.tsx",
@@ -1688,10 +1688,10 @@
         "name": "Partial<InputAdornmentProps>"
       }
     },
-    "KeyboardButtonProps": {
+    "OpenPickerButtonProps": {
       "defaultValue": null,
       "description": "Props to pass to keyboard adornment button",
-      "name": "KeyboardButtonProps",
+      "name": "OpenPickerButtonProps",
       "parent": {
         "fileName": "material-ui-pickers/lib/src/_shared/PureDateInput.tsx",
         "name": "DateInputProps"
@@ -2721,7 +2721,7 @@
     },
     "renderInput": {
       "defaultValue": null,
-      "description": "Override input component",
+      "description": "Render input component. Where `props` – [TextField](https://material-ui.com/api/text-field/#textfield-api) component props\n@example ```jsx\nrenderInput={props => <TextField {...props} />}\n````",
       "name": "renderInput",
       "parent": {
         "fileName": "material-ui-pickers/lib/src/_shared/PureDateInput.tsx",
@@ -2788,10 +2788,10 @@
         "name": "Partial<InputAdornmentProps>"
       }
     },
-    "KeyboardButtonProps": {
+    "OpenPickerButtonProps": {
       "defaultValue": null,
       "description": "Props to pass to keyboard adornment button",
-      "name": "KeyboardButtonProps",
+      "name": "OpenPickerButtonProps",
       "parent": {
         "fileName": "material-ui-pickers/lib/src/_shared/PureDateInput.tsx",
         "name": "DateInputProps"
@@ -3680,7 +3680,7 @@
     },
     "renderInput": {
       "defaultValue": null,
-      "description": "Override input component",
+      "description": "Render input component. Where `props` – [TextField](https://material-ui.com/api/text-field/#textfield-api) component props\n@example ```jsx\nrenderInput={props => <TextField {...props} />}\n````",
       "name": "renderInput",
       "parent": {
         "fileName": "material-ui-pickers/lib/src/_shared/PureDateInput.tsx",
@@ -3747,10 +3747,10 @@
         "name": "Partial<InputAdornmentProps>"
       }
     },
-    "KeyboardButtonProps": {
+    "OpenPickerButtonProps": {
       "defaultValue": null,
       "description": "Props to pass to keyboard adornment button",
-      "name": "KeyboardButtonProps",
+      "name": "OpenPickerButtonProps",
       "parent": {
         "fileName": "material-ui-pickers/lib/src/_shared/PureDateInput.tsx",
         "name": "DateInputProps"
