+++ lib/src/_shared/PureDateInput.tsx
@@ -8,7 +8,7 @@ import { useUtils, MuiPickersAdapter } from './hooks/useUtils';
 import { InputAdornmentProps } from '@material-ui/core/InputAdornment';
 import { getDisplayDate, getTextFieldAriaText } from '../_helpers/text-field-helper';
 
-type MuiTextFieldProps = TextFieldProps | Omit<TextFieldProps, 'variant'>;
+export type MuiTextFieldProps = TextFieldProps | Omit<TextFieldProps, 'variant'>;
 
 export interface DateInputProps<TInputValue = ParsableDate, TDateValue = MaterialUiPickersDate> {
   open: boolean;
@@ -25,7 +25,12 @@ export interface DateInputProps<TInputValue = ParsableDate, TDateValue = Materia
   TextFieldProps?: Partial<MuiTextFieldProps>;
   // ?? TODO when it will be possible to display "empty" date in datepicker use it instead of ignoring invalid inputs
   ignoreInvalidInputs?: boolean;
-  /** Override input component */
+  /**
+   * Render input component. Where `props` – [TextField](https://material-ui.com/api/text-field/#textfield-api) component props
+   * @example ```jsx
+   * renderInput={props => <TextField {...props} />}
+   * ````
+   */
   renderInput: (props: MuiTextFieldProps) => React.ReactElement;
   /**
    * Message displaying in read-only text field when null passed
@@ -52,7 +57,7 @@ export interface DateInputProps<TInputValue = ParsableDate, TDateValue = Materia
    * Props to pass to keyboard adornment button
    * @type {Partial<IconButtonProps>}
    */
-  KeyboardButtonProps?: Partial<IconButtonProps>;
+  OpenPickerButtonProps?: Partial<IconButtonProps>;
   /** Custom formatter to be passed into Rifm component */
   rifmFormatter?: (str: string) => string;
   /**
